Lesson 13/requirements.txt
asgiref==3.10.0
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
